move auth to own module
