+++ oura/__init__.py
@@ -9,5 +9,6 @@ Oura API Library
 It's a description for __init__.py, innit.
 
 """
-from .client import OuraClient, OuraOAuth2Client
+from .auth import OAuthRequestHandler, OuraOAuth2Client, PersonalRequestHandler
+from .client import OuraClient
 from .client_pandas import OuraClientDataFrame
